requirements.txt
fastapi
uvicorn
pydantic
requests
trafilatura
beautifulsoup4
pytest
python-multipart
google-generativeai
python-dotenv
playwright
playwright-stealth
